requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
greenlet==3.0.3
playwright==1.42.0
ply==3.11
pyee==11.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
soupsieve==2.5
typing_extensions==4.10.0
